Fix thumbnails date/time
Use EXIF_TAG_DATE_TIME_ORIGINAL instead of EXIF_TAG_DATE_TIME which
seems more reliable and is part of EXIF rather than IFD 0.

EXIF_TAG_DATE_TIME is the "Date and Time" field and is way off in some
of my pictures.

EXIF_TAG_DATE_TIME_ORIGINAL is the "Date and Time (original)" field
which seems to be reliably correct.

Also initialise the date[] buffer to "\0" so that if for some reason
there is no date/time then none will be displayed rather than a repeat
of the last one. This will make it easier to identify any problem there.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>