Use scandir(3) to get the list of images
Previously we were using readdir(3) to look for the images to use in the
album. The main problem with this is that it reads the entries in
directory order, while this mostly worked OK, I did notice a couple of
images were appearing out of sequence in the album.

We can use scandir instead and make use of the filter and compar
functions to have it filter out non-images and sort the list in name
order.

For the filter function we hook up a modified version of is_image_file()
This will filter out any non-images. In is_image_file we also replace
rindex(3) with strrchr(3) as rindex has been removed from POSIX.1-2008.

For the compar function we use alphasort(3) which will sort the entries
in name order.

Sorting the images by name should generally work. Sorting by mtime may
also be useful and could be added as an option later.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>