Completely switch to VIPS
This expands on the previous commit, b25dd3a ("Use VIPS to determine
image orientation"), and uses VIPS for the image resizing also.

Using the 'lanczos' interpolator as we were with GraphicsMagick gives us
the following results processing 213 images

    47.57user 3.19system 0:38.46elapsed 131%CPU

whereas with GM we were getting

    106.73user 4.14system 0:40.23elapsed 275%CPU

so a bit faster with much reduced CPU usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>